requirements.txt
Django>=3.1,<3.3
djangorestframework>=3.12,<3.13
psycopg2-binary>=2.8.6
python-decouple>=3.4,<3.5
django-cors-headers>=4.3.1,<4.4
drf-yasg>=1.21.7,<1.22